test(cuda): Catch error from find on non-NixOS
The CUDA script would fail on the last `find | nawk` command when:

1. `/dev/nvidia0` is present
2. `ldconfig` returns no matching libs
3. `/run/opengl-driver` doesn't exist (ie. not NixOS)

This caused the `find` to fail on a non-existent directory and caused
activations to fail because it runs the script with `set -e`.

It's reproduced by not automatically creating the directory for every
test in the suite. It's now only created for the two tests that exercise
NixOS behaviour. It's fixed by catching the error from `find` and
returning an empty result like we did previously for `ldconfig` in
5c3cf9b1.

Failure with the change to the test and not the script:

    ✗ cuda disabled when nvidia0 device present but libcuba absent [1327]
       tags: cuda
       (from function `assert_success' in file /nix/store/612z53vfi1py1qf2cii0nzrqcigfgbag-bats-with-libraries-1.11.1/share/bats/bats-assert/src/assert_success.bash, line 42,
        in test file cuda.bats, line 73)
         `assert_success' failed
    …
    ++ /nix/store/vsia946b1b3bqz3j9cznlibc5l834n5f-findutils-4.10.0/bin/find /var/folders/8q/spckhr654cv4xrcv0fxsrlvc0000gn/T/nix-shell.ub27qI/bats-run-cmTBoC/test/2/project-2/fake_fhs_root/run/opengl-driver -type f -print
       ++ /nix/store/328armcjagvyjz2iczccz9jn3s3zb0g2-nawk-20250116/bin/nawk -v 'lib_pattern=lib(cuda|nvidia|dxcore).*\.so.*' 'BEGIN {files=""} $0 ~ lib_pattern {files = ( files == "" ? $NF : files ":" $NF)} END {print files}'
       + SYSTEM_LIBS=

I think this script is still a good candidate for rewriting in Rust
because we can be more explicit about handling exit codes and error
output from individual commands (currently we have to swallow because
everything ends up in the activation), there's less chance of common
quoting and piping mistakes that we get from shell script, and it's
possible to test each part independently. We should consider when to
prioritise that separate from this ticket though, he says hoping that
this is the last of these bugs for now.