fix: stop using disallowedReferences for manifest builds
As observed in #3051 and discussed in #3020, our use of disallowedReferences
was causing build failures based on otherwise-immaterial references from the
log and buildCache outputs. This update replaces our use of that feature with
an external check performed after the "successful" build, one advantage of
which is that we can now display the exact "why-depends" reason for failing
the build, rather than just reporting a failure with no way to examine the
failed package to see what went wrong.