feat: make flox-build.mk consistent across build types
This patch updates the build logic and target names to be more consistent
across the sandbox "pure", sandbox "off", and NEF build cases. It does this
by better leveraging the COMMON_BUILD_VARS_template, using consistent target
filenames for each build type, and by instituting the following dependency
flow for all builds:

- NEF:$($(pvarname)_evalJSON) or Manifest:$(buildScript)
  - $($(pvarname)_buildJSON)
    - $(pvarname)_CHECK_RESULT_LINKS
      - $(pvarname)_CHECK_BUILD
        - $($(pvarname)_buildMetaJSON)
          - $(BUILD_RESULT_FILE)

Interspersed amongst the various build types there are type-specific actions
to be performed for the harvesting of logs, calculating extra build deps, etc.,
but the consistency that comes from imposing this basic flow makes it much
easier to make changes over time.

This change also introduces the initial check-build-prerequisites target as a
real file with a timestamp to force the re-evaluation of all subsequent targets
in the build graph.

In code review it was surfaced that there is a tiny race condition between the
evaluation of a NEF build and its subsequent build. We view this as an extremely
unlikely situation given the way that Make will schedule these sequentially, but
have added a test and error message if the drv file should disappear between the
evaluation of these two targets.

Other issues addressed:

- includes the fix to include NEF inputs in the output closure for a build (#3102)
- includes NEF inputs in the closure calculations when flagging unwanted requisites
  in the build closure
- updates the parsing of the NEF "reflect.targets" output to reflect that it no
  longer contains NEF build output names
- introduces the '$(_V_)' decorator to each of the clean targets so that we can
  resume displaying that output from the CLI
- fixes bug in the clean target to clean all files associated with a build by
  grouping them all in a common directory according to $(_pname)