refactor: move `LockedUrlInfo` to `build` provider
Besides likely being due for further refactoring,
we want to prepare for passing custom nixpkgs urls to the expression build subsystem.
Since expression builds require a base nixpkgs as well (determining that is TBD), the notion of locked urls from which a flakeref can be derived is required within the domain of the build.
To avoid a circular dependency on publish, moving `LockedUrlInfo` to builds as a first step seems sensible.
However, the format may change as we figure out how to retrieve these locked urls dynamically.