feat(build): dynamically define nixpkgs flakeref of expression builds
Allow passing a nixpkgs flakeref to the build subsystem for use as

1. the provider of `nixpkgs.lib` for the nef library
2. the base package set for expression builds

Primarily `publish` requires the locked url of a build to be truthful.
We ensure this by deriving a flakeref for the expression build from the `LockedUrlInfo`.
Currently we still hardcode a specific revision of nixpkgs, for builds and publishes of expression builds.
In the future we will extend the catalog server to provide relevant locked urls
as using nix to derive this proved inefficient.