fix: make consistent layered activation in both sandbox modes
Update pure manifest build to perform the layered dev + build environment
activation with the exact same arguments (in the same order) as done in the
sandbox="off" case. This includes the use of the "build" mode, which leaves
the "build-sandbox" mode unused, so this commit removes support for that mode
as well.

Updated comments regarding the move from `env -i` to the `env-filter` script.