fix: display only first 3 "extra" packages in build closure
When flagging "extra" packages not found in the "runtime-packages" closure
it's possible for that to be a long list, and when that happens it's rarely
useful to display screens of output when many of the packages will have been
flagged on account of the same root cause.

This patch limits the number of `nix why-depends` invocations to 3, while
reporting the total number of "extra" packages identified in the build closure.