fix(activate): Fix error from _profile_scripts
I noticed this error while looking at reporting activation errors with
traps for #3101. The `-d ''` argument specified that we expected a NUL
char after the space delimited values but the command we run never
provides one:

    read [-ers] [-a aname] [-d delim] [-i text] [-n nchars] [-N nchars] [-p prompt] [-t timeout] [-u fd] [name ...]
        …
        -d delim
                The first character of delim is used to terminate the input line, rather than newline.  If delim is the empty string, read will terminate a line when it reads a NUL character.

Before:

    % bash -xc 'declare -a _var; read -r -d "" -a _var < <(echo foo bar); echo "${_var[@]}"'
    + declare -a _var
    + read -r -d '' -a _var
    ++ echo foo bar
    + echo foo bar
    foo bar

    % bash -exc 'declare -a _var; read -r -d "" -a _var < <(echo foo bar); echo "${_var[@]}"'
    + declare -a _var
    + read -r -d '' -a _var
    ++ echo foo bar

After:

    % bash -exc 'declare -a _var; read -r -a _var < <(echo foo bar); echo "${_var[@]}"'
    + declare -a _var
    + read -r -a _var
    ++ echo foo bar
    + echo foo bar
    foo bar

The `nullglob` option also does not produce a NUL char:

    % bash -c 'shopt -s nullglob; echo *.nomatch | cat -ev'
    $

There are other ways that we could write this function but then we have
to do more work to restore `nullglob` so it doesn't seem worth it.