refactor(publish): Use open_path()
Partial revert of 7b990aa7 which made `EnvironmentPointer::open` public.

The typo of "environment" originally caught my attention and then it
occurred to me that the `unreachable!()` message was inverted because it
would only be triggered when opening a non-path environment.

We can use `open_path()` instead, as already used by `providers::build`,
and then just call trait methods on `Environment` without having to
explicitly convert to `ConcreteEnviroment::Path` or similar.

Track the project path in `CheckedEnvironmentMetadata`, rather than the
`.flox` path and working backwards, to make this easier. This doesn't
lessen our guarantees about what's in the tracked repo because we're
still opening an environment.