fix: include `env-filter` in flox-package-builder
The `env-filter` script was being built and installed as a single-file package,
and this commit updates the build to instead include it as a script in the
libexec subdirectory of the flox-package-builder package.

We did this after observing problems with the `env-filter` script not being
found in CI, and in one instance we also observed it was not marked executable.