test: weaken nested activation PATH assertion
Somewhere in `PATH` there is a `::`, and it turns out that shells often
expand this to `:.:` e.g. they turn a `::` into a path component that
represents the current directory. The nested activation tests rely on
knowing the explicit `PATH` before and after the nested activation so
that we can ensure that the original `PATH` hasn't been somehow mangled
or dropped.

I only observed this behavior with integration tests run against a
Nix-built flox binary, and only for Fish and tcsh.

Since the shell is doing this `::` -> `:.:` expansion, we can no longer
rely on knowing the exact value of `PATH`. To make the tests pass, we'll
make the assertion weaker by only asserting that `PATH` is prefixed by
the Flox environment directories and `before_path:`, which is explicitly
prepended to `PATH` by the test helpers. The idea is that if we see
`before_path`, we're seeing *all* of `PATH`, and we don't need to check
the exact value of `PATH`.