feat(containerize): Allow --runtime and --file
Allow both `--runtime` and `--file` arguments to be specified rather
than being mutually exclusive, so that macOS users with both `docker`
and `podman` on their PATH can choose which proxy runtime they want to
use, whilst still being able to write to a file outside the proxy.

I'm not convinced that `--file` is actually that useful since we added
the ability to write to runtime registries but it's hard work to get
usage metrics and removing it would be a breaking change.

The new logic is:

1 No args
  a. Detect proxy runtime, and
  b. Detect output runtime or write to default file
2. --runtime foo
  a. Use foo for the proxy and output runtime
3. --file bar
  a. Detect proxy runtime and write to file bar
4. --runtime foo --file bar
  a. Use foo for the proxy runtime and write to file bar

I considered putting the `output` match statement as a method on
`OutputTarget` so that it could be unit tested but I don't think it
provides much value given the readability of the statement, the coverage
we have from integration tests, and it's slightly harder to reason about
the args outside the context of the handler.