feat(build): pass a nixpkgs url for dependency expression builds
In order to provide better compatibility with manifest builds,
we need to allow building nix expressions with a compatible set of nixpkgs.
Generally this will the "latest available" set of nixpkgs used when building,
nix expressions can't be expected to be compatible.
Instead we will use the nixpkgs revision referenced by the toplevel group (if available).