chore: make nef nixpkgs evaluations more permissive
Allow evaluating (and building) unfree and insecure packages.
It's the responsibility of the catalog to deny _installation_ downstream.
For builds, provide a more user friendly, i.e. permissive experience.

This keeps denying broken and unsupported packages,
as building or using those is unlikely
to be resulting in a working build.