fix(pull/push): Better current_dir error
We got a Sentry panic from `flox pull`:

- https://flox-dev.sentry.io/issues/6648555991

`current_dir` can return an error if the dir was deleted or you no
longer have permissions to access it. The former can be reproduced with:

    % mkdir tmp && cd tmp && rmdir ../tmp && flox pull dcarley/tailscale

    thread 'main' panicked at flox/src/commands/pull.rs:109:78:
    called `Result::unwrap()` on an `Err` value: Os { code: 2, kind: NotFound, message: "No such file or directory" }
    note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

Now it returns:

    % mkdir tmp && cd tmp && rmdir ../tmp && flox pull dcarley/tailscale
    ❌ ERROR: Failed to get current directory: No such file or directory (os error 2)

`into_dir` needed moving up because of the partially borrowed value.
The two match statements could have been combined but the resulting code
was less readable.

I also fixed `flox push` which had the same issue.