fix(python): use poetry activation command
We use the `poetry env info --path` command to get the location of the
Python virtual environment built by Poetry. However, if Poetry fails to
find a suitable Python version (e.g. one that matches a version
requirement in `pyproject.toml`) activation will fail and the output of
this command will be empty.

Poetry has its own activation command that we can `eval` instead, and
this command displays an error when one is encountered rather than
silently failing.