chore: bump testing nixpkgs URL
Bump testing nixpkgs URL to stay in sync with the nixpkgs we use to
build Flox.

If we don't do this, it will cause build tests to fail, as certain build
tests currently depend on the fact that patchShebang will patch using a
bash that is in the closure of flox-interpreter. If the nixpkgs used for
patchShebang is different than the one used to build flox-interpreter,
the build will fail with errors like:
```
❌ 1 packages found in /nix/store/sg8yy9rf2wi4jakxjdfq7ly220647wqh-mypkg-1.0.2a
   -      not found in /nix/store/4w3fvpfl2yjabdb536ij74zlp8hwwz1l-environment-build-mypkg
```