feat: improve eval times of expression builds under version control
Currently, the nef libraries reference `pkg-dir` as a "raw path" (and implicit `impure-eval`)
As such, nix will copy _any path referenced by nix expressions_ to the store unfiltered.
That is bad for at least two reasons:
1. otherwise ignored files, possibly containing secrets are copied into the world-readable `/nix/store`
2. any build artifaccts are copied as well, which can be very large wasting space and time it takes to copy them

This commit adds a `git-subdir` argument to the `nef` entrypoint.
When that argument is provided, we expect `pkgs-dir` to be the path to a git repository root,
and `git-subdir` to be the relative path to the expression dir within that repository.
Accompanying are changes to `flox-build.mk` to split NIX_EXPRESSION_DIR into these two components,
and provide nix arguments accordingly.

Note: the implementation may be a bit. of a hack but provided a quick performance boost.
That said, for the sake of complexity this might be refactored in a future change.