fix(main): Quieten warning about effective UID
Change the warning to debug because it seems alarmist when we don't
appear to have any reports in GitHub or Slack of bugs that are caused by
`USER` being different from the effective username. It will still be
included with `-vv` output though, so we'll see it when asking for debug
output in issues.

It's common for `sudo` to be configured to preserve `$HOME` from the
calling user but I'm not clear about the conditions where `$USER` would
be different from the effective user. Based on this old thread about the
original logic in `flox-bash` it seems like this might not be needed
anymore:

- https://flox-dev.slack.com/archives/C054SAD6NLD/p1687900692516919

But I think that's best combined with fixing `sudo` behaviour in:

- https://github.com/flox/flox/issues/3025

So I've also added a debug trace to indicate if and when we are actually
resetting the `USER` and `HOME` environment variables in the meantime.