feat: poll for publish completion
Uses the new `/store_path/status` endpoint to poll for the completion
of the publish process. The publish process doesn't end with uploading
the artifacts, since there is a signing process and propagation of
metadata to the catalog server. If something fails after uploading
artifacts, we need to inform the user so that they don't mistakenly
think that the publish was a success.