feat: comment out options.systems in template
This comments out `options.systems` in the manifest template. When all
of the systems are present, it's functionally equivalent to omitting
this option entirely, so we might as well omit it. This has the added
benefit of not generating a composition merge notification when one
environment includes another without otherwise overriding anything.