fix: populate $(BUILD_OUTPUTS) prior to constructing manifest build DAG
Issue #3217 documents an instance where the manifest build script preparation
failed to substitute a package reference prior to invoking the build. Turns out
this was caused by a combination of the order of processing and not attempting
substitution for all known builds, and splitting the processing of all known
$(MANIFEST_BUILDS) into two passes addresses the problem.

Also updated existing test to confirm ability to build dependency containing
multiple dashes, as we initially thought that was the origin of the bug.

Closes #3217.