fix: prepend newline when passing through Nix error
Since we're passing the error through without any modification and it
may be multiline, put it on a newline.

For a `flox install` this changes the error from
```
❌ ERROR: Failed to build environment:

Failed to link environment: error: don't have permission to lock Nix database in '/nix/var/nix/db'.
       Nix may be running as non-root in a single-user installation,
       or the Nix daemon may have crashed
```

To
```
❌ ERROR: Failed to build environment:

Failed to link environment:
error: don't have permission to lock Nix database in '/nix/var/nix/db'.
       Nix may be running as non-root in a single-user installation,
       or the Nix daemon may have crashed
```