fix: always prepend newly-activated environments to PYTHONPATH
The previous logic made the mistake of appending new environment paths
to the PYTHONPATH. This patch fixes that by updating the logic to
instead prepend new entries to the PYTHONPATH.

The previous version also failed to re-order the PYTHONPATH in the event
that the newly-activated environment path was already present in the
string. That was not correct either because the most recently-activated
environment should always be first in the path, regardless if it was
already present in the string. This patch updates the logic to prepend,
then remove duplicate elements from the path in every instance which has
the effect or re-ordering the PYTHONPATH correctly.

Updated existing bats python activation test to exercise both the
expected prepending to an existing PYTHONPATH and the removal of
duplicates from the resulting PYTHONPATH.