fix: service-level variables not applied
You can set environment variables at both the environment level (via the
`[vars]` section) or the service level via
`services.<name>.vars.<name>`. Previously the service level variables
were not applied. There are two bugs involved.

The first bug is caused by the fact that `process-compose` has an
internal templating engine that allows interpolating variables defined
in the `vars` section of the config file. These variables are _not_
environment variables, and are only used for templating purposes.
However, the manifest uses `vars` as the name of the environment
variables section, and so when the service configuration is serialized
to YAML we end up translating environment variables to template
variables. This explains why it was possible to use template syntax by
mistake in the service configuration.

The second bug is that by default, which is an extremely questionable
decision, `process-compose` will expand shell variables in the config
file _at load time_. This means that if the command for your service
is `echo "hello $user"`, the value of `$user` will be looked up in the
environment of `process-compose` (or an empty string if it's not found)
and substituted into the config file as a pre-processing step before
running your command. This means that if `$user` isn't present in the
environment, the command used to spawn your service actually becomes
`echo "hello "`. Disabling this behavior is done via the top level
`disable_env_expansion` option.