feat: remove layered activation from manifest build
Our aim in performing a manifest build is to replicate as closely as
possible the experience of running those same build script commands
from within an interactive `flox activate -m dev` shell (i.e. using
the "develop" environment).

But the manifest build differs from the interactive case as follows:

1. strives to achieve reproducibility by redacting environment variables
   that are otherwise managed by the Flox environment
2. prevents references to the "develop" environment from being added to
   the resulting build

The first of these is straightforward, using `env -u` to redact environment
variables as we kick off a build, and our approach to the second is the
following:

1. create "build-wrapper" environments for each build, customized to only
   include packages as found in the "runtime-packages" manifest attribute
2. perform the build using the "develop" environment
3. replace all references to the "develop" environment path with that of
   the "build-wrapper" environment

The easiest and most reliable way to replace those references (i.e. without
resorting to the use of tools like patchelf) is by way of a binary string
substitution, which carries the requirement that the paths involved in the
substitution must be of the same length. In other words, the path to the
"develop" environment _used for the build_ must have the same length as the
path to the "build-wrapper" environment.

We accomplish this by performing the build using a _copy_ of the "develop"
environment found in a path with the same strlen as the "build-wrapper"
environment, so that following a successful build we can replace references
to the former with the latter as we copy the output to its final path.

As we implemented the above we noticed that we had been passing buildDeps
(the set of packages required for a build) as additional elements in the
buildInputs array, but that carries the implication that "file" inputs
are sourceable hooks, and our buildCache input definitely does not meet
that criteria. So this patch removes the buildDeps from buildInputs and
instead passes them by way of an additional buildDepInputs variable in
the derivation.

Also found that with this change in buildInputs handling that we need to
sign $out on darwin in all cases, not just for local builds.