feat: remove layered activation from manifest build
Our aim in performing a manifest build is to replicate as closely as
possible the experience of running those same build script commands
from within an interactive `flox activate -m dev` shell (i.e. using
the "develop" environment).

But the manifest build differs from the interactive case as follows:

1. strives to achieve reproducibility by redacting environment variables
   that are otherwise managed by the Flox environment
2. prevents references to the "develop" environment from being added to
   the resulting build

The first of these is straightforward, using `env -u` to redact environment
variables as we kick off a build, and the second is addressed by introducing
a level of indirection:

1. create symlink from $out/nix-support/FLOX_ENV -> "develop"
2. perform build using $out/nix-support/FLOX_ENV, being careful to
   redact Flox-managed variables from the process's environment before we start
3. replace $out/nix-support/FLOX_ENV with link to "build-wrapper"