fix(publish): Improve copy outputs debugging
Improve debug output for copying outputs and explicitly skip copying
when there are no outputs. This was already a noop when
`upload_build_outputs` iterates over an empty vec.

This helped me to debug a test where it was copying to the wrong path
because of a replayed mock from a previous test invocation and when some
handcrafted metadata didn't contain any outputs at all.