fix: detect build deps
We make an attempt to detect build dependencies written with the
typical Nix syntax in the user's manifest builds. However, the
comment in this diff breaks the command being run by the Makefile
because the previous line has a line continuation.

Without this fix, the Go pure build example was not able to complete
a multi-stage build where the first stage vendors dependencies and the
second stage uses the output of that build.