feat: add generic path fixing command
We already have functionality for prepending and deduplicating
environment directories as we add them to a `PATH`-like variable. This
change just exposes that functionalty as a generic command that prints
out the modified `PATH`-like variable. This means that it's up to the
caller to properly _set_ a variable to this output and then export it.
This is done on purpose to keep the command generic.