fix: add version.command test, amended test function
Went to add a version.command test for evaluating `$()` expansions within
commands and found that the test function that runs `nix derivation show`
was relying on experimental features being enabled on a system-wide basis.
This patch both adds that flag to the `nix` invocation and exposes stderr
produced by the test so that errors from the command are visible to the
developer.