fix: don't run on-activate hook for build activation
The refactoring of the layered build activation exposed that we were
relying on the redacting of env variables between the outer and inner
activations for hiding the _effects_ of the on-activate hook, but that
the hook was still being invoked by the outer activation.

This patch addresses the original intent which was to avoid running the
on-activate hook in the first place for build activations.