fix: correctly detect mock regen
Previously we were preferring the presence of the response file when
determining whether to re-run a job and not honoring the
`skip_if_output_exists` field in all cases. Now we use this field to
completely override which path to check when determining whether to
re-run a job.

This change addresses a bug in which deleting an `env/*` directory
wouldn't regenerate the mock if the corresponding `*.yaml` file still
existed.