feat: outdent scripts read from manifest
Outdent `profile`, `hook` and `build` scripts.
In the manifest.toml, its a reasonable stylistic choice to indent scripts for readbility, e.g.

```toml
-- 8< --

[profile]
common = '''
    echo <<EOF

        An important paragraph of text
      that spreads multiple
      lines.

    EOF
'''
```

However the toml spec treats the string contents literally, meaning the indent is part of the content that gets locked.
Until now we have written files including the literal content found in the lockfile, meaning we created shell files that may be partially indented.
This inadvertently breaks HEREDOCS for example.

In response we will now outdent the content in the lockfiles by the minimal common indentation within the script.