test(build): Add tests for closure check
Which can be triggered both with `runtime-packages` and with packages in
a non-toplevel group. For the latter we needed an additional mock
because the package groups are included in the request body which is
matched when finding the appropriate mock.

I found it useful to regex match some of the error message that's
produced by `flox-build.mk` by to verify that we're formatting it
correctly but I don't think it's worth the effort of matching all of the
`why-depends` output because it's very variable.