feat(build/publish): Remove feature flags
In preparation for 1.5 release.

We aren't using these anywhere else in code or tests, via the struct or
environment variables, so there is nothing else to change. The
`doc-build` directory mentioned in the ticket was already merged in
c9d0d169.

I've left the feature flag on `upload` because we're not exposing this
to users and perhaps(?) it will be removed in the near future.