feat(template): Add build to manifest template
I've taken the shortest but complete example I can from the concept docs
because we don't want to clutter up everyone's environments with a
really long command.

It's a little odd to have `[build]` and `[build.myproject]` next to each
other when uncommented but it both matches the style of the other
headings and makes the project specific build easier to read.

I've signposted the build and publish commands so that users know what
to do next but there's not much more to say about them like we do for
services beacuse they're pretty self explanatory.