feat: add ENOENT detection following #3257
PR#3257 removed layered activations from the manifest build, resulting in a
much better chance of success because it more closely emulates what a user
can do on the command line, but this also introduced the possibility that a
file that is present in the build-time [copy of the] "develop" environment
may not be present once that has been substituted with the "build wrapper"
environment controlled by the runtime-packages attribute.

This patch adds a new `validate-build` script to perform all build checks,
which includes moving the existing closure checks out of `flox-build.mk`.
It also introduces a new "missing files in build output" check which inspects
the built package looking for path strings that refer to the [final] "build
wrapper" environment after substitution, and then asserts the existence of
each path found. Any path that is not found is then cross-referenced with
the "develop" environment, and if found there then the script prints a message
suggesting that the user add that package to the `runtime-packages` attribute.