fix: don't allow building from a managed environment
Builds are necessarily associated with source code, so in general we're
expecting builds to be performed using Path environments committed to
git.

We disallow publishing from a ManagedEnvironment, so we should do the
same for builds for consistency.

If we change our mind, it will be easier to allow building from
ManagedEnvironment's than disallowing.