fix: test hyphenated package name, remove whitespace
Found bug where test was referring to `$(_pvarname)` instead of `$(_pname)`
which meant the test was expecting the pname with dashes replaced with
underscores. Updated the package name in the test to "my-package" so that
we can detect that mistake if ever we make it again.

Also removed whitespace logic from test that is no longer necessary after
splitting out the logic into the seperate validate-build script.