fix: Nix patch for CVE-46415 and CVE-46416
Apply nixpkgs patch for Nix 2.28 to our build of Nix to fix CVE-46415
and CVE-46416.

I considered pulling in a second nixpkgs flake or overriding Nix src,
but just applying the patch seemed the most straightforward since we're
already building Nix from source, and we want as minimal of a change as
possible since we're cutting a release.