fix: correctly set start-services-var
Flox documentation states: "true" if this activation started
services, "false" otherwise.  The previous logic exposed the state of
the process-compose process, but that is more of an internal detail and
should be irrelevant to the user. On a host where lots of service
related environments are in play, this would always be false.  Where the
intent appears to be if the activation started "your" services or not.