fix: show build output during publish
flox#3236 changed the behavior of `flox publish` to store the build
output so that it can be shown when the build fails. However, this means
that a user running `flox publish` sees no output during the build which
they may not know is happening. This change restores the build output so
that users don't get the wrong impression that something is frozen
while the build is happening.

As part of this change, since the build output is shown during the
build, it is no longer collected and printed in case of an error.
Instead, the output is shown during the build and if there is a failure
a short message containing the exit status is shown.