fix(catalog): Trace headers when metrics enabled
Move the generation of trace headers to the generated client methods
rather than setting them statically when the client is initialized.

This fixes two issues:

1. Trace headers were never being added because there was no active span
   when `init_catalog_client` is called.
2. Even if they were, the trace ID would have been correct but the span
   ID would not have, causing the request to appear at the wrong point
   in the waterfall.

There's still a chance that the span ID might be incorrect if we're
using `Span::enter()` rather than `#[instrument]`, `.instrument()`, or
`Span::in_scope()` across async boundaries but that's not as important
as having no headers and we can fix those later if needed.

In addition to the new unit tests I've also observed the correct request
headers when using `mitmproxy` in intercept+block mode, with both:

1. `FLOX_DISABLE_METRICS=false FLOX_SENTRY_DSN=<DSN> flow show hello`
2. `FLOX_DISABLE_METRICS=true FLOX_SENTRY_DSN=<DSN> flow show hello`

Once we've added integrations to our other services this will allow us
to see trace waterfalls from the CLI, `catalog-server`, and `publisher`
for users that have metrics enabled, which will help us diagnose
performance issues.